#14 Added Timetravel support
Except for its children, MapContainer props are immutable: changing them after they have been set a first time will have no effect on the Map instance or its container.

That is why passing a state to the center prop of MapContainer does not automatically make it stateul. That is why Timetravel also does not change the map display when the location in the state does change.

https://stackoverflow.com/questions/64665827/react-leaflet-center-attribute-does-not-change-when-the-center-state-changes#answer-64667351